Introduce reference counting for multiple alloca with same address
In cases where the runtime detects "reallocation" of memory, this
allocation is not added to the stack variables list. Nevertheless
the full number of allocations is freed at the end of the block.

This happens if the optimizer decides to use a single allocation
for different sub blocks in the current block.