Implement detection of array cookies
Instead of collecting all bitcasts and filtering them, we observe that
in case a GEP instruction is applied to an allocation it usually is used
to calculate the pointer of the actual array.
In case two GEP instructions are present, the first is used to add
necessary padding to the array cookie and it's bitcast instructions can
be safely ignored.