add array cookies to MallocData
Introduces a new data structure for array cookie information that stores the
store instruction that writes to the cookie and the value of the padding
passed to the GEP instruction that computes the array pointer passed to the
user.

This commit also introduces several asserts to catch cases which are not
expected and should be further investigated.