share typeart-no-heap and typeart-alloca args
This makes these two command line arguments available in both
the MemInstFinderPass and TypeARTPass. This allows the
MemInstFinder to only collect the necessary instructions. As
heap allocations are only collected on unoptimized code this
in turn allows for a stricter analysis of the given code.

To share the command line arguments, the cl::opt instances
have been shared between the two passes. As the TypeARTPass
depends on the MemInstFinderPass to already be loaded, the
instances have been moved to the MemInstFinderPass and are
now referenced with external linkage in the TypeARTPass.