add a tracing id to the logs for every MPI call
Instead of repeating the general information
about each call on each logging output, it's
now printed once in the beginning alongside
an id which can be used to identify subsequent
logs for that call. This way, only the rank
and id have to be printed for each logging
output to be uniquely assignable to one
log of the general information.