move data query to struct creation
Currently some data is queried at the time of checking a type.
This commit moves this data to the data structures and their
query to the creation function of these structures. This way,
an error returned by any of the type checking functions will
always be an actual type error.