check ambiguous types for pointers to structs
When given a pointer to a struct, the pointer could be meant
to have the type of the struct itself or it's first member,
given the first member has an offset of 0.
With this commit, the first member of a struct is now
recursively considered a candidate for the type check
if the struct type itself does not match.