add a --command option to typeart-templ.sh.in
Currently only the name of the executable file that is compiled by
the script can be specified which is then also used for executing it.
With this change, a command option can be specified which defaults
to the name of the executable but can be set to e.g. call mpirun.