add a --types_yaml parameter to typeart-templ.sh.in
This parameter allows to explicitly specify the location of the
types.yaml file generated by typeart.
Also, by default it is no longer set to 'types.yaml' but to the
basename of the source filepath with a '.types.yaml' extension.
As this is a full path instead of just a filename this makes it
easier to reason about problems with the generation or loading
of the file as it no longer depends on the current working
directory.