only print errors if the overall typecheck failed
In cases where a typecheck first checks a type against a whole
struct and then against the first member of the struct, an
error is printed for the first case. This makes actual errors
harder to identify.
With this patch, all errors occuring during the typecheck are
printed a traces and an error is only printed if the typecheck
fails overall.